+++ examples/count_curveto_segments.rs
@@ -1,29 +1,19 @@
-extern crate svgdom;
-
 use std::env;
-use std::io::Read;
-use std::fs::File;
+use std::fs;
 
-use svgdom::{ AttributeId, AttributeValue, Document, ElementId, FilterSvg, PathCommand };
+use svgdom::{AttributeId, AttributeValue, Document, ElementId, FilterSvg, PathCommand};
 
-fn main() {
+fn main() -> Result<(), Box<std::error::Error>> {
     let args: Vec<_> = env::args().collect();
-
     if args.len() != 2 {
         println!("Usage:\n\tcount_curveto_segments in.svg");
-        return;
+        std::process::exit(1);
     }
 
-    let mut file = File::open(&args[1]).unwrap();
-    let length = file.metadata().unwrap().len() as usize;
-
-    let mut input_data = String::with_capacity(length + 1);
-    file.read_to_string(&mut input_data).unwrap();
-
+    let input_data = fs::read_to_string(&args[1])?;
     let doc = Document::from_str(&input_data).unwrap();
 
     let mut count = 0;
-
     for (id, node) in doc.root().descendants().svg() {
         if id == ElementId::Path {
             let attrs = node.attributes();
@@ -34,4 +24,6 @@ fn main() {
     }
 
     println!("This file contains {} CurveTo segments.", count);
+
+    Ok(())
 }