+++ src/attribute_value.rs
@@ -8,7 +8,7 @@
 
 use std::fmt;
 
-use {
+use crate::{
     Angle,
     AspectRatio,
     AttributeId,