+++ src/node.rs
@@ -9,7 +9,7 @@
 use std::iter::FilterMap;
 use std::cell::{Ref, RefMut};
 
-use {
+use crate::{
     tree,
     Attribute,
     AttributeId,