+++ src/lib.rs
@@ -60,7 +60,7 @@ doc for details.
 
 */
 
-#![doc(html_root_url = "https://docs.rs/svgdom/0.16.1")]
+#![doc(html_root_url = "https://docs.rs/svgdom/0.17.0")]
 
 #![forbid(unsafe_code)]
 #![warn(missing_docs)]