+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Parse feComposite's `k1`, `k2`, `k3` and `k4` attributes.
+- Parse `startOffset` attribute.
 
 ## [0.16.1] - 2019-03-15
 ### Changed