+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.17.0] - 2019-06-12
 ### Added
 - Parse feComposite's `k1`, `k2`, `k3` and `k4` attributes.
 - Parse `startOffset` attribute.
@@ -498,7 +500,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/RazrFalcon/svgdom/compare/v0.16.1...HEAD
+[Unreleased]: https://github.com/RazrFalcon/svgdom/compare/v0.17.0...HEAD
+[0.17.0]: https://github.com/RazrFalcon/svgdom/compare/v0.16.1...v0.17.0
 [0.16.1]: https://github.com/RazrFalcon/svgdom/compare/v0.16.0...v0.16.1
 [0.16.0]: https://github.com/RazrFalcon/svgdom/compare/v0.15.0...v0.16.0
 [0.15.0]: https://github.com/RazrFalcon/svgdom/compare/v0.14.0...v0.15.0